Gallery add support for singleton embedded images
Added support for singleton galleries, so you can embed multiple images into an article for example. Clicking on an image brings up the mobile carousel where user can navigate to view all the images in the article. Each embedded image is created as a separate gallery, but they are associated by using the same id.